gitaly/hooks: Properly configure the logging infrastructure
Our "gitaly hooks" subcommand does not properly initialize the logging
infrastructure even though it has the actual Gitaly configuration at its
fingertips. The only thing the configuration does is to set up the log
level of the gRPC logger, and the fact that this does indeed work is
only because the server code is replacing the gRPC logger in an `init()`
function.

Let's properly set up the configuration via `gitalylog.Configure()`.
While this does not yet fix the fragile gRPC logger setup, this will be
handled in the next commit where we move the setup of the gRPC logger
into `Configure()` itself. Furthermore, this also fixes inconsistent
logging formats and levels that previously completely disregarded what
the administrator had configured.