log: Simplify configuration of the logging infrasturcture
The way to configure our logging infrastructure is quite weird as we
rely on a set of global variables defined in "internal/log", only to
pass them back into the package anyway.

Refactor the code to not do that so that we can make the loggers
private. This also requires us to extend the signature of the
configuration function to accept an `io.Writer` and the logging hooks,
which have previously been configured in other packages by accessing
those global variables.