golangci-lint: Drop current forbidigo rules
We have some forbidigo rules in place that disallow the usage some
context functions in our tests. This is realized by excluding all files
from the rule that do not end in "_test.go". Unfortunately though, this
keeps us from using the linter for actual production code, which is
arguably more important.

As we cannot both have our cake and eat it, too, let's drop our current
rules. We will add new rules that apply to production code in the next
commit.