featureflag: Remove the `CatfileBatchCommand` flag
The `CatfileBatchCommand` flag was used to rollout the `--batch-command`
option of `git-cat-file(1)`. Since the flag has been enabled for a while
now, we can delete the flag and default to the enabled state.

We cannot remove the old code because we are still dependent on the
minimum git version supporting `--batch-command`, which is scheduled to
land in git 2.42.0.