grpc: Convert to use `logrus.FieldLogger` interface
Convert the dnsresolver, backchannel and sidechannel packages to accept
logging interfaces instead of an explicit `logrus.Entry`. Later down the
road we will provide our own interfaces to abstract the `logrus` package
away from our own internal interfaces, so this change is an early step
into that direction.

Note that this requires a change in the backchannel package to convert
the field logger to an `logrus.Entry` by calling `WithField()` first.
We do this by adding a "component" field that points to backchannel,
which has the added benefit that we can now more clearly identify when
log messages come from inside that subsystem.

Furthermore note that we need to change the logging middleware tests to
not import our public "client" package anymore and instead use our
internal one. This change is required due to a new cyclic dependency
between the "internal/log" tests and the "client" package.