ci: Provide common dictionary of component versions
We specify about a dozen different versions in our CI definitions, and
many of these versions are repeated in multiple locations. At times,
this makes it hard to update one of these versions while making sure
that all of the places have been updated that should be.

Unfortunately, it is not possible to have YAML anchors that refer to a
single plain string, only. That is, the following is not possible:

```yaml
.versions:
    git: "v2.41.0" &version_git
```

We are lucky though, because at GitLab we have noticed that this is
indeed an issue and have introduced the `!reference` extension that is
available in GitLab 13.9 and later. So let's unify all our versions into
a common dictionary so that we can easily reference them. This has two
major advantages, namely that all used versions are defined in a single
spot and that the versions are deduplicated.