log: Move hook logger initialization to its only user
The gitaly-hooks command has somewhat special logging requirements as
they can neither log to stdout (as this would be used by Git), nor can
they log to stderr (as this would be shown to the user). We thus either
log into a file "gitaly_hooks.log", or alternatively just discard all
log entries altogether.

While this setup logic is thus closely tied to the gitaly-hooks logger,
the setup is contained in the "internal/log" package. Let's move it to
its only user to make things more self-contained and not clutter the
generic log package with highly specialized functionality.