cmd/gitaly: Properly configure the logger for the "check" subcommand
Same as with the preceding commit, the "check" subcommand does not
properly initialize its logging infrastructure. Refactor the code to do
so.

Note that we also refactor the code such that we first load the config
and initialize the logging infrastructure before we start to perform the
checks. This allows us to exit early in case the configuration does not
exist or cannot be parsed, whereas until now we already started printing
status information to the process's standard output.