Limiter: Add concurrency in_progress log field
The concurrency limiter attaches some log fields to gRPC logs. Those
fields include useful information for operations. There is one log field
called `limiter.concurrency_queue_length`. It tells the number of
requests in the queue. That log field consists of in-progress requests
and in-queue requests. Although the number of in-queue requests can be
calculated from the concurrency configuration and that queue length,
it's troublesome to do so.

This commit adds `limit.concurrency_in_progress` log field for the
number of concurrent requests.