Support running Gitaly with transactions enabled
This commit adds a configuration option for running Gitaly with
transaction management enabled. While transactions are not yet ready
for production use, this allows them to be configured for development
and testing purposes.

Transactions are integrated with rest of Gitaly by configuring an
interceptor in GitalyServerFactory that handles beginning,
committing and rolling back transactions for RPCs.

PartitionManager is initialized on Gitaly's start up, and closed when
Gitaly is terminating. Initializing and stopping the partition manager
opens and closes the database handles for each storage where Gitaly
stores some transaction processing related state.