Pass repository path directly in GetCustomHooks
GetCustomHooks is currently taking a repository and a locator as a
parameter to construct the repository path where to get the custom
hooks from. This make the function less flexible as it assumes the
caller has a locator or a configuration object to construct a locator
from available. It would be better if the function was just concerned
with extracting the hooks from a given repository rather than also
locating the repository. This commit changes the function to take a
repository path in directly so its more flexible.