Limiter: Fix concurrency queue_length log semantic
Recently, the concurrency queue_length log field includes both
in-progress requests and waiting requests. This semantic reflects the
internal implementation of the limiter. The limiter uses two systems:
queue tokens and concurrency tokens. A request must acquire the queue
token before acquiring the concurrency token.

This semantic is confusing. In fact, Prometheus metrics differentiate
those two pieces of information. This commit detaches the queue length
and the in-progress logs.