cli/praefect: Do not require logger when reading the config
We currently require a logger when reading the configuration. This is
because `getConfig()` may print a warning when the election strategy is
misconfigured. This creates a chicken-and-egg problem though, because we
can't properly configure the logger yet until we have read the config.

This warning message really only matters to the "serve" subcommand, as
all the other client-side commands would not use the election strategy
anyway. So let's pull out the warning into the `run()` function that is
used to back the "serve" subcommand, which allows us to only log that
warning after we have configured the logger.

Reading the configuration does not require a logger anymore, so we can
remove the parameter. While at it, the function is also renamed to
`readConfig()` such that it indicates better what it is actually doing.