Rearrange state directory's contents
TransactionManager's state directory can be simplified somewhat.

Since the state is no longer inside the repository itself, we can
remove the unnecessary 'wal' subdirectory that holds the other
directories. This was mostly in place to collect all of the other
directories in a single subdirectory in the repository.

That leaves 'packs' and 'hooks'. The content of the 'hooks' directory
never was really about the write-ahead log. The protobuf marshaled
log entries carry the logged hooks directly themselves. The hooks in
the directory were for supporting MVVC in order to retain multiple
versions of the hooks for use by transactions executing at different
log indexes.

The state in the 'packs' directory is directly related to the WAL
though. The packs in the directory are the logged pack files that are
part of the log entries. Let's thus rename the 'packs' directory to 'wal'.
This also leaves space in the name for logging other state in the directory,
not just packs. For example, if we wanted to log the hooks through the
directory as well as opposed to slurping them in memory as we do now.