storagemgr: Remove packed-refs lockfile on start
Should a process which has locked the `packed-refs` file crash, the
transaction manager may fail to apply a log entry due to the stale
lockfiles.

To prevent this, we remove both `packed-refs.lock` and `packed-refs.new`
if present when initializing the transaction manager. We provide no
grace period for these files as we know that no other process should be
touching the repository when the transaction manager is starting.

Changelog: added