log: Return errors when logger configuration fails
When configuration of the logger fails then we print a log message via
the logrus package. This has the problem that this logger is not yet
properly configured and thus may or may not be the correct thing to do.

Refactor the code to instead return an error when configuration of the
logger fails. This lets callers handle the error appropriately by
themselves.