operations: Drop useless split index test for UserSquash
In the past, we used to set the `core.splitIndex` configuration in some
of our RPCs. This past is long gone, and the chance for that behaviour
to reoccur "by accident" is... quite slim. We still have a test though
that verifies whether that setting is set or not, which seems kind of
pointless nowadays.

Remove that test.