gittest: Detect default object format via environment variable
Right now Gitaly still has two different build variants for its tests,
one for each of the supported object formats. This also means that we
have to essentially build the code base twice, doubling the amount of
data we need to put into our caches and increasing build times overall.

Drop the build flag in favor of an environment variable that modifies
the default object format.