Disable RemoveAll tests with WAL
RemoveAll is a problematic RPC as it empties out the entire storage.
Doing so, it removes the disk state Gitaly is still operating on. This
causes issues with WAL as the transaction manager and the database don't
expect their disk state to suddenly disappear. Disable RemoveAll tests
with WAL until the RPC has been properly integrated.