praefect: Rename 'repository' to 'relative-path'
Historically the Praefect tracking subcommands have not distinguished
between the relative path[0] used by clients and the replica path used
internally by Praefect[1]. Now that replica paths are in common usage we
need to make it clear which path the user is specifying.

In preparation for adding a new flag for replica paths, rename the
`-repository` flag to `-relative-path`, which matches the names used in
our documentation.

[0] https://docs.gitlab.com/ee/administration/gitaly/#client-generated-replica-paths
[1] https://docs.gitlab.com/ee/administration/gitaly/#praefect-generated-replica-paths-gitlab-150-and-later