praefect: Use replica paths in track-repository
The `track-repository` subcommand has assumed that the relative path and
replica path of a repository are the same. However, since v15.0 this is
no longer the case and the command has been unable to handle paths that
use the `@cluster` format.

To resolve this, differentiate between relative paths, those used by
clients (e.g. rails), and replica paths used internally by Praefect and
require both to be passed by users.

We temporarily manually set replica path to match relative path in
`track-repositories` to keep tests passing. In the following commit this
will be updated to use the deserialized value from the user.