log: Work around race when replacing gRPC logger
The `log.Configure()` function is expected to be called exactly once,
either in the `main()` function or in the respective subcommand. This is
because we set up global resources, namely our normal loggers as well as
the gRPC logger. Consequentially, when multiple tests call that function
in parallel, we run into a data race that gets detected by Go's race
detector.

While our normal loggers will eventually be converted to be non-global,
the gRPC logger cannot ever be non-global because the gRPC library does
not provide a way to do so. We thus introduce an ugly workaround in the
form of a new variable that asks the `log` package to skip replacing the
gRPC logger, which is only set in our `testhelper` package.

While not pretty, it at least addresses the data race.