tests: Do not use `log.Default()` logger
Do not use the `log.Default()` logger in our tests. We typically don't
want to see log outputs on neither standard output nor on standard
error -- if any of the tests wants that, they should instead be passing
in an explicitly configured logger.

Instead, we convert tests to use a discarding log entry.