gitaly/config: Refactor how warnings are generated
We're printing some logs when validating the Gitaly configuration, which
results in a chicken-and-egg situation: we shouldn't be using the logger
yet as it's not yet configured, but in order to configure it we need the
logger as part of the configuration validation.

We have two different log messages which we're printing:

    - We print a debug entry for every configured directory. This is
      weird by itself alreday and doesn't seem to serve much of a
      purpose. Furthermore, as the logger is not yet configured, its log
      level would default to "info" and thus the messages would not be
      printed anyway. This message is thus removed without replacement.

    - We print a warning when "transitioning" is enabled. This one is a
      bit weird because we always print this warning when validating,
      but in fact this warning is only relevant for the "serve" command
      and not for clients. This warning is thus moved into our server
      code.

With those two changes we don't need to log anything in our config code
anymore.