client: Use discarding logger for DNS resolver by default
The `DefaultDNSResolverBuilderConfig()` constructs a default config that
is to be used by dependents on the Gitaly configuration. Part of this
default config is the logger, which we set to `log.Default()`. This is
arguably unexpected though, because the default logger will cause us to
print message to standard output.

Let's refactor the code to instead use a discarding logger by default.
If clients want to have logging for this code, they should instead
supply their own logger.