Fix flaky TestUpdater_contextCancellation
TestUpdater_contextCancellation is flaky as it's canceling the context
and expecting the next method call to fail with context cancellation
error. The context cancellation leads to the process being terminated
asynchronously so it's not guaranteed to be done yet by the time of
the next method call. Keep calling until the context cancellation
error is returned to fix the flake.