gitlab: Don't store mocked functions in global variables
In the GitLab mock client we have a bunch of mock functions that can be
passed to `NewMockClient()`. These functions are assigned to global
variables without any reason whatsoever -- if a caller wants a different
set of functions they would pass something else to the constructor.

Remove the variables and declare the mocks as plain old functions.