repository: Fix ReplicateRepository test depending on host umask
One of our tests for ReplicateRepository implicitly depends on the host
umask. This is becaues we have a hardcoded vote that expects a specific
vote for the hooks vote, but that vote does in fact include permission
bits that change depending on the current umask.

Fix this by dynamically detecting the expecte mode bits such that we can
compute the expected vote instead of having to hardcode it.