Add a recording logger for tests
Our tests are currently largely discarding all logs. This is done
as we run our tests verbosely in the CI and the log output exceeded
what the CI could at maximum display. Discarding the logs can  make
it more difficult to determine why the test exactly failed. This is
especially true when concurrency is involved and a background worker
not directly tied to the request fails. While the background worker
would log the reason it failed, this is not visible in the tests.

This commit solves the problem by adding a new logger that records
its output and only prints it out if the test failed. This way in
the general case we still have no log output in the tests but we
will have logs when the test fails and we'd benefit from them.

Ideally we'll standardise all test to use this logger and prevent
usage of other loggers in tests.