Replace use of NewDiscardingLogEntry with the recording logger
Gitaly now has a helper for recording log output and outputting it
only in tests. Replace the usage of the discarding logger with the
recording logger. These changes are search and replace so these call
sites didn't even need a log entry to begin with but rather use a
FieldLogger.

NewDiscardingLogEntry is removed as it is no longer used.