service/hook: Fix pack-objects hook tests with runtime dir
In 2559c4b9d (Remove gitaly_pack_objects_limiting_remote_ip feature
flag, 2023-05-12), we have removed a feature flag for pack-objects
limiting by remote IP. In the process, we have refactored some tests for
running the pack-objects hook with or without a runtime directory. This
refactoring broke the test though because `runTestsWithRuntimeDir()` is
quite misleadingly named: it doesn't run tests, but instead returns a
function that will run the tests for you. That function is not being
executed anymore though, and thus the test basically does nothing.

Refactor the function to actually run tests when invoked. As the test is
now essentially broken with the SHA256 object format because it still
uses a seed repository, we also convert it to generate test data at
runtime.

Note that we also immediately quarantine the test given that it quite
flaky nowadays. This will need to be fixed in a follow-up change.