Implement a method for test case shared loggers
We were previously discarding logs in most of the test cases. As
such, the loggers are not being passed around in the test cases
and multiple loggers are instantiated. The log messages printed
to multiple loggers are unordered which makes it more difficult
to see the exact output of the logs. As a workaround, this commit
implements a sharing mechanism to ensure all logs in a given test
case end up in the same logger. The first call to the SharedLogger
method creates a new logger, and the follow up calls will use the
same logger within the same test case.

The function will be wired in a follow up commit.