Print Gitaly and Praefect logs on test failure
Gitaly and Praefect logs are currently redirected to files. To debug
test failures, one has to open the file separately and download the
log artifacts from CI. Instead of redirecting the logs to a different
files, use the recording logger and output the logs automatically when
a test case fails.