gittest: Dynamically detect whether to skip SHA256 tests
We are now in a state where all functionality except for backups work
with the SHA256 object format. This means we can also get rid of the
build tag that we used for our tests in favor of dynamically skipping
over tests when we see that the object format is unsupported.

Introduce a new function `SkipWithSHA256()` and convert all tests using
the `!gitaly_test_sha256` build tag to use it instead.