Cgroup: Add an option to include Gitaly process in Cgroup hierarchy
When cgroup feature is enabled in Gitaly, all children Git commands are
assigned with a corresponding per-repository cgroup. All of those
cgroups are also under the control of a parent cgroup. This hierarchy
limits the resource usage of those commands and prevents them from
dominating the node resources.

In an incident, we discovered that the main Gitaly process is not a part
of that cgroup hierarchy. It's fine most of the time because the Gitaly
process itself is not resource-hungry. Unfortunately, in the incident,
the Gitaly process leaks memory and grows its memory usage gradually.
After a certain point, the process becomes unresponsive. This situation
cannot be self-resolved without human intervention.

This commit adds an "IncludeGitalyProcess" config to include the Gitaly
process in the parent cgroup.

When the IncludeGitalyProcess configuration is enabled, the main cgroup
process's PID is added to Gitaly's cgroup hierarchy. The process doesn't
possess its own exclusive limit. Rather, it shares the limit with all
spawned Git processes, which are also managed by the repository cgroup,
under the umbrella of the parent cgroup.

This configuration offers superior protection and flexibility compared
to setting a fixed limit specifically for the Gitaly process. The Gitaly
process typically has more resource leeway than the Git processes and is
likely to be terminated last in the event of a problem.

For https://gitlab.com/gitlab-org/gitaly/-/issues/5535