issue_templates: Make feature flag rollout a multi issue template
Rolling out feature flag is generally done over multiple issues. Most
often the work can be stretched for 3 releases:
1. Deploy code with feature flag <Release X>
2. Enable feature flag on staging <Release X>
3. Enable feature flag on production <Release X>
4. Default enable the feature flag <Release X+1>
5. Remove feature flag <Release X+2>

Currently we create a single issue from the `Feature Flag Roll Out`
template and use the same issue to track work over the releases. This is
problematic since:
1. Most of the time the issue is blocked waiting for a release to
happen.
2. Often we prematurely close the issue without removal of feature flag,
which means a lot of flags remain as is.

To solve both these problems, let's split the `Feature Flag Roll Out`
issue template into three:
1. `Feature Flag Roll Out`
2. `Feature Flag Default Enable`
3. `Feature Flag Removal`
with this, the first issue can be closed in <Release X> after creating
the required subsequent issues. This ensures we have discrete work which
can be tracked easily and also ensures that cleanup tasks are created in
our backlog making it easier to track/delete old flags.