gitaly/config: Allow custom hooks directory to be unset
While we allow custom hooks directories to not exist, we do not
explicitly handle the case where the global custom hooks directory is
not set. And in fact, this case doesn't happen right now because if
`[hooks] custom_hooks_directory` is unset, we will set it to a value
derived from `[gitlab-shell] dir` instead. And as that value is enforced
to never be empty, we know that the custom hooks directory will always
be set.

We are about to change this though by allowing `[gitlab-shell] dir` to
be unset. This would cause us to resolve the global custom hooks
directory to a relative path though, which is of course not what we
want.

Explicitly handle the case where the custom hooks directory is unset so
that we properly skip it.