gitaly/config: Verify explicitly configured custom hooks directory
We do not check whether an explicitly configured custom hooks directory
actually exists on disk. This makes the administrator blind to any kind
of misconfiguration in case they for example have a typo in the path.
Let's start to validate that the directory exists if the value was not
derived from the `gitaly-shell.dir` configuration.