gitaly/config: Allow configuration without `gitaly-hooks` directory
Our configuration validation logic enforces that the `gitaly-hooks`
directory must be set. This directory is used for two different things,
both of which already have proper replacements:

    - It identifies the location of the `.gitlab_shell_secret` file,
      which is required in order to connect to GitLab. This can also be
      configured via `[gitlab] secret` and `[gitlab] secret_file`.

    - It identifies the location of custom hooks. This can also be
      configured via `[hooks] custom_hooks_directory`.

The `gitaly-hooks` section can thus be considered legacy and should
ideally not be used anymore.

Rewrite the configuration validation to allow the `gitaly-hooks`
directory to remain unset. Instead, we will verify that the GitLab
secret is properly configured.